+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Updated table styles to simplify and unify table margins
 - Implemented Search for Students table
+- Implemented Search for Groups table
 
 ## 0.15.1
 - Updated views for new grade structure